pyproject.toml
[project]
name = "resustack-ai-service"
version = "0.1.0"
description = "AI Service for Resustack - Resume Builder with AI Coach"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.124.2",
    "langchain>=1.1.3",
    "langchain-anthropic>=1.2.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "tiktoken>=0.12.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.0",
    "pre-commit>=4.5.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.8",
]

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py312"
src = ["backend"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# MyPy configuration
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
strict = false
plugins = ["pydantic.mypy"]
mypy_path = "backend"

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["backend"]
addopts = "-v --tb=short"

# Development commands:
# - Dev server: make dev (or: uv run python -m backend.main)
# - Tests: uv run pytest
# - Lint: uv run ruff check backend tests
# - Format: uv run ruff format backend tests
# - Type check: uv run mypy backend
